kernel: proc: modify scheduler policy
Those changes should only affect queues 7 ~ 14.

We add a new flag to the `proc` structure to indicate whether the process
has been scheduled. This allows us to track each queue's current process.

We also change the scheduler's placement policy so that now it inserts the
process at the end of the queue. This is done to avoid starvation.